requirements.txt
Django==4.2.0
djangorestframework==3.14.0
drf-spectacular==0.26.0
